// DashboardApplication/admin-dashboard/src/app/generated/helpers/interface.generated.ts
export interface BaseHttpResponse<T> {
	status: boolean;
	message?: string;
	totalCount?: number;
	data?: T;
}